requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
packaging==24.0
python-dotenv==1.0.1
pytz==2024.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
